pyspark_app/requirements.txt
pywebhdfs
mysql-connector
kubernetes
kafka-python
sqlalchemy
pandas
PyMySQL